DATAMONGO-1609 - Fix failing tests.
Fix issues pointed out by failing tests. Main focus was to restore functionality and not a Java 8 code cleanup. So, this one still needs some love and polishing.

+++ spring-data-mongodb/src/test/java/org/springframework/data/mongodb/core/query/IsTextQuery.java
@@ -103,7 +103,7 @@ public class IsTextQuery<T extends Query> extends IsQuery<T> {
 	}
 
 	@Override
-	public IsQuery<T> skippig(int skip) {
+	public IsQuery<T> skippig(long skip) {
 
 		super.skippig(skip);
 		return this;
